Minor cast fixes
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@38855 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSArray.m
@@ -455,7 +455,7 @@ static SEL	rlSel;
 	    {
 	      NSString	*key;
 
-	      key = [NSString stringWithFormat: @"NS.object.%u", i];
+	      key = [NSString stringWithFormat: @"NS.object.%lu", (unsigned long)i];
 	      [(NSKeyedArchiver*)aCoder encodeObject: [self objectAtIndex: i]
 					      forKey: key];
 	    }
@@ -724,14 +724,14 @@ static SEL	rlSel;
 	  id		val;
 
 	  array = [NSMutableArray arrayWithCapacity: 2];
-	  key = [NSString stringWithFormat: @"NS.object.%u", i];
+	  key = [NSString stringWithFormat: @"NS.object.%lu", (unsigned long)i];
 	  val = [(NSKeyedUnarchiver*)aCoder decodeObjectForKey: key];
 
 	  while (val != nil)
 	    {
 	      [array addObject: val];
 	      i++;
-	      key = [NSString stringWithFormat: @"NS.object.%u", i];
+	      key = [NSString stringWithFormat: @"NS.object.%lu", (unsigned long)i];
 	      val = [(NSKeyedUnarchiver*)aCoder decodeObjectForKey: key];
 	    }
 	}
